pyproject.toml
[project]
name = "wine-quality-mlops-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "dvc>=3.66.1",
    "fastapi>=0.135.1",
    "hydra-core>=1.3.2",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "mlflow>=3.10.1",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
]
